% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_scarcity_model.R
\name{create_water_scarcity_model}
\alias{create_water_scarcity_model}
\title{Create a model estimating the relationship between days without clean water
and water infrastructure age}
\usage{
create_water_scarcity_model(study_data)
}
\arguments{
\item{study_data}{A data frame with the needed predictor variables}
}
\value{
A model to predict future clean water availability given census block attributes
}
\description{
Create a model estimating the relationship between days without clean water
and water infrastructure age
}
