% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_suitability.R
\name{determine_site_suitability}
\alias{determine_site_suitability}
\title{Determine site suitability for}
\usage{
determine_site_suitability(study_data, value_function_config,
  mental_models)
}
\description{
Determine site suitability for
}
