% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_model.R
\name{load_climate_scenario}
\alias{load_climate_scenario}
\title{Load a climate scenario csv given at a particular path}
\usage{
load_climate_scenario(path)
}
\arguments{
\item{path}{Path to the file}
}
\value{
climate scenario data.frame
}
\description{
Load a climate scenario csv given at a particular path
}
