% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megadaptr.R
\docType{package}
\name{megadaptr-package}
\alias{megadaptr}
\alias{megadaptr-package}
\title{Megadapt spatial microsimulation}
\description{
A spatial microsimulation model of residents,
    administrators and climate for evaluating fresh and storm
    water investment policies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cpritcha/megadaptr}
  \item Report bugs at \url{https://github.com/cpritcha/megadaptr/issues}
}

}
\author{
\strong{Maintainer}: Calvin Pritchard \email{calvin.pritchard@asu.edu}

Authors:
\itemize{
  \item Andres Baeza
  \item Fidel Serrano Candela
  \item Manuela Vanegas Ferro
  \item Sean Bergin
}

}
