% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_model.R
\name{update_climate}
\alias{update_climate}
\title{Generates yearly precipitation and runoff (in millimeters) by census block}
\usage{
update_climate(study_data, next_year_study_data, climate_scenario)
}
\arguments{
\item{study_data}{A data frame of study area data}

\item{climate_scenario}{A data frame containing yearly precipitation and runoff amounts (in millimetres)
per census block}
}
\value{
yearly runoff and precipitation values by census block
}
\description{
Generates yearly precipitation and runoff (in millimeters) by census block
}
