% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponding_model.R
\name{update_ponding}
\alias{update_ponding}
\title{Generate ponding data}
\usage{
update_ponding(study_data, ponding_models)
}
\arguments{
\item{ponding_models}{A list of ponding models for each region}
}
\value{
ponding
}
\description{
Generate ponding data
}
