% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_scarcity_model.R
\name{update_water_scarcity}
\alias{update_water_scarcity}
\title{Generate random water scarcity data for a year for all census blocks}
\usage{
update_water_scarcity(water_scarcity_model, study_data, week_of_year)
}
\arguments{
\item{water_scarcity_model}{A model capable of predicting water scarcity probabilities by census block}

\item{study_data}{Census block data
\describe{
\item{antiguedad_dist}{Fresh water infrastructure age in years}
\item{critic_z}{Indicator variable of whether census block has less than 4 hours of water per day}
\item{falta_dist}{Number of houses in census block without water}
}}

\item{week_of_year}{The number of weeks up until the present week of the current year}
}
\value{
data frame with pk and cumulative number of days without clean water this week and year by census block
}
\description{
Generate random water scarcity data for a year for all census blocks
}
